refactor: move from_config.rs to inside from_config module
