--- src/blueprint/from_config_cp/mod.rs
@@ -1,3 +1,0 @@
-mod server;
-
-pub use server::*;
