+++ src/blueprint/mod.rs
@@ -1,10 +1,9 @@
 mod blueprint;
 mod compress;
-mod from_config;
 mod into_schema;
 mod timeout;
-mod from_config_cp;
+mod from_config;
 
 pub use blueprint::*;
 pub use timeout::GlobalTimeout;
-pub use from_config_cp::*;
+pub use from_config::*;
